Module Lin.MakeDomThr

A functor to create Domain and Thread test setups. We use it below, but it can also be used independently

Parameters

module Spec : CmdSpec

Signature

val interp_plain : Spec.t -> Spec.cmd list -> (Spec.cmd * Spec.res) list

val interp : Spec.t -> Spec.cmd list -> (Spec.cmd * Spec.res) list

val interp_thread : Spec.t -> Spec.cmd list -> (Spec.cmd * Spec.res) list

val gen_cmds : int -> Spec.cmd list QCheck.Gen.t

A fueled command list generator.

val gen_cmds_size : int QCheck.Gen.t -> Spec.cmd list QCheck.Gen.t

val shrink_triple : (Spec.cmd list * Spec.cmd list * Spec.cmd list) -> (Spec.cmd list * Spec.cmd list * Spec.cmd list) QCheck.Iter.t

val arb_cmds_par : int -> int -> (Spec.cmd list * Spec.cmd list * Spec.cmd list) QCheck.arbitrary

val check_seq_cons : (Spec.cmd * Spec.res) list -> (Spec.cmd * Spec.res) list -> (Spec.cmd * Spec.res) list -> Spec.t -> Spec.cmd list -> bool

val lin_prop_domain : (Spec.cmd list * Spec.cmd list * Spec.cmd list) -> bool

val lin_prop_thread : (Spec.cmd list * Spec.cmd list * Spec.cmd list) -> bool