Module Lin_effect

module Make_internal (Spec : Lin.Internal.CmdSpec) : sig ... end

Functor to build an internal module representing Stdlib.Effect-based tests

val fork : (unit -> unit) -> unit

Helper function to fork a process in the underlying Stdlib.Effect-based scheduler

val yield : unit -> unit

Helper function to yield control in the underlying Stdlib.Effect-based scheduler

module Make (Spec : Lin.Spec) : sig ... end

Functor to build a module for Effect-based testing