Module Lin_thread.Make

functor to build a module for concurrent testing

alert experimental This module is experimental: It may fail to trigger concurrency issues that are present.

Parameters

module Spec : Lin.Spec

Signature

val lin_test : count:int -> name:string -> QCheck.Test.t

lin_test ~count:c ~name:n builds a concurrent test with the name n that iterates c times. The test fails if one of the generated programs is not sequentially consistent. In that case it fails, and prints a reduced counter example.

val neg_lin_test : count:int -> name:string -> QCheck.Test.t

neg_lin_test ~count:c ~name:n builds a negative concurrent test with the name n that iterates c times. The test fails if no counter example is found, and succeeds if a counter example is indeed found, and prints it afterwards.