Module Lin_thread.Make_internal

Functor to build an internal module representing concurrent tests

alert internal This module is exposed for internal uses only, its API may change at any time

Parameters

module Spec : Lin.Internal.CmdSpec

Signature

val arb_cmds_triple : int -> int -> (Spec.cmd list * Spec.cmd list * Spec.cmd list) QCheck.arbitrary

val lin_prop : (Spec.cmd list * Spec.cmd list * Spec.cmd list) -> bool

val lin_test : count:int -> name:string -> QCheck.Test.t

val neg_lin_test : count:int -> name:string -> QCheck.Test.t